add types to syscall errors
The underlying error types are `repr(transparent)`, so they'll work fine
in the FFI boundary. but the types ensure we don't misinterpret them.

In the future, we can move to adding more types like:

- Block IDs.
- Boolean? But we can't do that unless we use 0 and 1 (negative 1 isn't
  valid).
- ActorID

Although many of these can just be type _aliases_. I used a transparent
tuple type here for extra safety so we don't misinterpret it.