feat: split the machine into an executor and a machine
The executor drives message execution on a given machine. By splitting
these types up, we can avoid type cycles _and_ avoid having to tell the
machine which kernel to use every time it executes a message.

From a type standpoint, the Kernel contains a CallManager which contains
a Machine. There are no cycles here.

From a flow standpoint, the machine (now executor) calls the call
manager which calls the kernel. This reverse flow is fine as long as it
doesn't show up in the type signatures.

Unfortunately... That reverse flow _does_ show up in type signatures the
second we try to make things generic:

1. The DefaultCallManager's type needed to know which Kernel to use.
2. The DefaultKernel needed to know which CallManager it contains.

The solution is to _not_ have the DefaultCallManger know which kernel to
use _statically_, but feed this information through when we actually
_call_ it. That keeps this information out of the type signatures.

The minimal solution was to change `execute_message` to take a kernel
factory on every call. Instead, I split it out into an Executor. The
executor is allowed to "know" about the kernel because the kernel
doesn't know about the executor.

The new type reference graph is (no cycles):

```
Executor -> Kernel -> CallManager -> Machine.
```