remove Blockstore child trait from Externs.
Also remove the blockstore/cgo requirement for the
fvm crate, which does not belong.

We ignore the BindSyscall rust doc example, since it
doesn't compile due to visibility issues.

Add consume() methods to StateTree and HAMT to consume
the instances and return the owned Blockstore. We need
to instantiate a working state tree because the Machine
loads it.

With these changes, we are now able to run cargo test
on the fvm crate.