executor: add raw_length argument.
The FVM takes the message to execute sans signature to minimize overhead.
Unfortunately, this means that we lose the real on-chain length of the message
as information. This is needed to calculate the message inclusion gas cost.

This commit introduces an argument in the Executor#execute_message method
for the node to report the original on-chain length of the message.

Additionally, it updates the conformance test vector runner
to account for the secp256k1 length bump, in equivalence to:

https://github.com/filecoin-project/lotus/blob/3fb71cdfea7399a3fa0033b48e0ea58e81ea1bd2/conformance/driver.go#L267-L275