fix: use a different CBOR library
Use a CBOR library that is tailored towards DAG-CBOR.

CIDs don't work as expected with Serde untagged enums, hence add a custom
implementation for the Pointer (de)serialization. That implementation works
similarly to the Serde untagged enum implementation.

First the data is deserialized into a common format that supports all types.
In our case that's `Ipld`, in Serde it's `Content`. Those are then converted
into the actual types the enum is using.

Directly converting from the serialized form into the concrete enum types is
not possible. One would need to try each of the possibilities sequentially
until one of them matches. The deserializer is consumed when deserializing,
hence it can only be used once. This is the reason to have the intermediate
step.