feat: add a feature to statically disable debugging in the sdk
When the debug feature is disabled, most actor debugging logic should
get statically compiled out.

I've left it enabled by default for now, but we should consider
disabling it by default for the final release. But if we do that, we
need _some_ way to re-execute actors with debugging enabled.