fix: make address networks global (#931)
We relied on `new_id` being "const" in several places, so the previous
patches weren't going to work. I had a few options:

1. Revert.

2. Make it a compile time option. This is strictly less flexible than
the current option.

3. Use a third value for "default" or "unset" that simply follows the
system default. This was probably a better solution than the existing
one, but made the Eq implementation a bit tricky.

So, instead of all of these solutions, I took a closer look at the issue:

1. Any solution that bundles "networks" with addresses will have a bit
of a race: any address constructed _before_ the network is set will be
wrong.

2. Addresses don't really _have_ networks. _Textual_ addresses do, but
addresses lose their networks as soon as they're parsed.

So I opted for a completely new option (that, honestly, more closely
matches lotus's behavior):

1. Remove networks from addresses.
2. Enforce that addresses have the correct network _on parse_.
3. Use the "current" network _on format_.

In this world, implementations should set the correct network before
trying to parse any addresses. However, if they get this wrong, they'll
get an actual _error_ instead of getting addresses that _should_ be
equivalent but aren't.

---

Breaking:

- Removes `network` from the `Address` type. I can't actually find any
calls in forest or the builtin actors outside of tests.
- Removes `set_network` from the `Address` type. Again, this wasn't
called outside of tests.
- Enforces that all textual addresses match the correct network on
parse. This is the change most likely to cause issues, if any.